Add files via upload
This project uses a linear regression model to predict house prices. After loading the data from a CSV file, it's split into features and the target variable, which is the sale price. The dataset is further divided into training and testing sets for model training and evaluation. The model is trained using the training data and then evaluated using the Root Mean Squared Error (RMSE) on both training and testing sets. Finally, a scatter plot is used to visualize the predicted versus actual house prices, helping to assess the model's performance.